+++ packages/react-components-lab/docs/CHANGELOG.md
@@ -1,3 +1,17 @@
+## @dhi/react-components-lab [1.13.1](https://github.com/DHI/react-components/compare/@dhi/react-components-lab@1.13.0...@dhi/react-components-lab@1.13.1) (2022-03-18)
+
+
+### Bug Fixes
+
+* **DashboardCard:** Export component from the library ([f58dbc0](https://github.com/DHI/react-components/commit/f58dbc03e1edb08ff34fed48d3163502af6951b7))
+
+# @dhi/react-components-lab [1.13.0](https://github.com/DHI/react-components/compare/@dhi/react-components-lab@1.12.6...@dhi/react-components-lab@1.13.0) (2022-03-18)
+
+
+### Features
+
+* **DashboardCard:** Add dashboardCard component ([1a6e48c](https://github.com/DHI/react-components/commit/1a6e48ccfe015f8abcf06443b3a62cb6b9f152ba))
+
 ## @dhi/react-components-lab [1.12.6](https://github.com/DHI/react-components/compare/@dhi/react-components-lab@1.12.5...@dhi/react-components-lab@1.12.6) (2022-02-28)